requirements.txt
pandas~=1.4.4
tqdm~=4.64.1
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-Caching==1.10.1
Flask-Limiter==2.7.0
Flask-Login==0.6.2
Flask-Mobility==1.1.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.0
python-dotenv==0.21.0
psycopg2==2.9.6
pyYAML==6.0
Werkzeug==2.2.2
gitpython==3.1.31